Decouples `field_box_t` from `pusher_box_t`
Every `pusher_box_t` needs a `field_box_t` to push its gyrons and,
previously, the latter object was built inside the factory responsible
to create the former. This arrangement forced every distinct pusher
object to have its *private* `field_box_t`, which results in duplication
if the electromagnetic field is the same for a significant collection of
gyrons. The problem is solved by decoupling the building of both objects
and passing a `field_box_t` pointer to the `pusher_box_t` factory. Also,
the burden of checking if the passed `field_box_t` objects are valid was
moved into the base `pusher_box_t` constructor, forcing it to be done
automatically for every derived class.